+++ Domain/CashDesk/IPaymentService.cs
@@ -2,7 +2,7 @@ namespace Domain.CashDesk;
 
 public interface IPaymentService
 {
-    Task PayCashAsync(int amount);
+    Task PayCashAsync(long amount);
     
-    Task PayCardAsync(int amount);
+    Task PayCardAsync(long amount);
 }
